+++ public/data/changelog.md
@@ -1,5 +1,9 @@
 ##### Dataset Changelog
 
+###### Oct 16, 2023
+
+- add EuroVis papers from 2022 to 2008
+
 ###### Oct 12, 2023
 
 - add ~30 authors who published in VIS 2022